More cleanup
Abstract away the notion of 'tracking orders' etc. behind some
utilities.

We may want to play with the interfaces further, esp. as we see people
build some strategies or write a few more of our own.

We probably also want to provide helpers for reacting to each new pool
state.